Rename component folders and files
