+++ src/pages/OfferPage/OfferPage.tsx
