+++ src/types/place-type.ts
@@ -1,0 +1,12 @@
+
+type PlaceType = {
+  name: string;
+  type: string;
+  imageUrl: string;
+  price: number;
+  rating: number;
+  isPremium: boolean;
+  inBookmarks: boolean;
+}
+
+export default PlaceType;
